% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrearVenn_function.R
\name{creaVenn}
\alias{creaVenn}
\title{Crea Venn Diagrams}
\usage{
creaVenn(area.vector = 0)
}
\arguments{
\item{area.vector}{Matriz de dos columnas, y con N renglones, donde N es el total de combinaciones menos uno}
}
\description{
Esta funcion permite crear Diagramas de Venn de 2 a 5 variables.
}
