% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrearVenn_function.R
\name{creaVenn}
\alias{creaVenn}
\title{Crea Venn Diagrams}
\usage{
creaVenn(combinaciones = 0, categorias = 0)
}
\arguments{
\item{combinaciones}{Matriz de dos columnas, y con N renglones, donde N es el total de combinaciones menos uno.}

\item{un}{vector del formato c('','') donde se escribe el nombre de las categorias en orden ABCDE}
}
\description{
Esta funcion permite crear Diagramas de Venn de 2 a 5 variables.
}
